requirements.txt
git+https://github.com/hroncok/easywebdav.git
Pillow==4.2.1
